+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGELOG
 ---
 
  * Add `StopWorkerExceptionInterface` and its implementation `StopWorkerException` to stop the worker.
- * Add the options `group` and `class-filter` to `FailedMessagesShowCommand`
+* Add the options `stats` and `class-filter` to `FailedMessagesShowCommand`
 
 5.3
 ---